+++ src/ios/CordovaLib/Classes/Public/CDVPlugin+Resources.h
@@ -17,20 +17,25 @@
  under the License.
  */
 
+#if !TARGET_OS_OSX
 #import <UIKit/UIKit.h>
-#import "CDVCommandDelegate.h"
+#endif
+#import "CDVPlugin.h"
 
-@class CDVViewController;
-@class CDVCommandQueue;
+@interface CDVPlugin (CDVPluginResources)
 
-@interface CDVCommandDelegateImpl : NSObject <CDVCommandDelegate>{
-    @private
-    __weak CDVViewController* _viewController;
-    NSRegularExpression* _callbackIdPattern;
-    @protected
-    __weak CDVCommandQueue* _commandQueue;
-    BOOL _delayResponses;
-}
-- (id)initWithViewController:(CDVViewController*)viewController;
-- (void)flushCommandQueueWithDelayedJs;
+/*
+ This will return the localized string for a key in a .bundle that is named the same as your class
+ For example, if your plugin class was called Foo, and you have a Spanish localized strings file, it will
+ try to load the desired key from Foo.bundle/es.lproj/Localizable.strings
+ */
+- (NSString*)pluginLocalizedString:(NSString*)key;
+
+/*
+ This will return the image for a name in a .bundle that is named the same as your class
+ For example, if your plugin class was called Foo, and you have an image called "bar",
+ it will try to load the image from Foo.bundle/bar.png (and appropriately named retina versions)
+ */
+- (UIImage*)pluginImageResource:(NSString*)name;
+
 @end