+++ src/ios/CordovaLib/Classes/Public/CDVPlugin+Resources.m
@@ -17,23 +17,22 @@
  under the License.
  */
 
-#import <UIKit/UIKit.h>
-#import "CDVPlugin.h"
+#import "CDVPlugin+Resources.h"
 
-@interface CDVPlugin (CDVPluginResources)
+@implementation CDVPlugin (CDVPluginResources)
 
-/*
- This will return the localized string for a key in a .bundle that is named the same as your class
- For example, if your plugin class was called Foo, and you have a Spanish localized strings file, it will
- try to load the desired key from Foo.bundle/es.lproj/Localizable.strings
- */
-- (NSString*)pluginLocalizedString:(NSString*)key;
+- (NSString*)pluginLocalizedString:(NSString*)key
+{
+    NSBundle* bundle = [NSBundle bundleWithPath:[[NSBundle mainBundle] pathForResource:NSStringFromClass([self class]) ofType:@"bundle"]];
 
-/*
- This will return the image for a name in a .bundle that is named the same as your class
- For example, if your plugin class was called Foo, and you have an image called "bar",
- it will try to load the image from Foo.bundle/bar.png (and appropriately named retina versions)
- */
-- (UIImage*)pluginImageResource:(NSString*)name;
+    return [bundle localizedStringForKey:(key) value:nil table:nil];
+}
+
+- (UIImage*)pluginImageResource:(NSString*)name
+{
+    NSString* resourceIdentifier = [NSString stringWithFormat:@"%@.bundle/%@", NSStringFromClass([self class]), name];
+
+    return [UIImage imageNamed:resourceIdentifier];
+}
 
 @end